feat(reader): 添加PDF注释功能和搜索高亮支持
- 实现PDF文本内容搜索覆盖层渲染
- 添加注释创建、导出、过滤和管理功能
- 支持EPUB章节片段锚点滚动定位
- 添加PDF缩放后重用文本内容搜索覆盖层
- 实现书签当前页面状态标记功能
- 添加注释类型、标签过滤器和批量操作
- 实现注释持久化存储和重启恢复
- 添加注释导出时Markdown敏感内容转义
- 实现大型注释侧边栏虚拟滚动优化
- 修复EPUB无搜索或文本注释时避免重复解析

+++ src-tauri/src/pdfkit.rs
@@ -56,7 +56,7 @@ mod platform {
             return Err(ENCRYPTED_DOCUMENT_ERROR.to_string());
         }
 
-        let page_count = unsafe { document.pageCount() } as usize;
+        let page_count = unsafe { document.pageCount() };
         if page_number > page_count {
             return Err("Invalid PDF page".to_string());
         }
